requirements.txt
python-dotenv
langchain
langchain-community
Pyrebase
pinecone-client
langchain_ollama
pycryptodome
flask_restful
setuptools
google-api-python-client